+++ src/tlc_ultralytics/engine/model.py
@@ -110,8 +110,6 @@ class YOLO(YOLOBase):
         if settings is None:
             settings = Settings()
 
-        settings.verify(training=False)
-
         if not settings.run_description:
             settings.run_description = DEFAULT_COLLECT_RUN_DESCRIPTION